+++ src/lib.rs
@@ -328,12 +328,13 @@ fn expand_connect(group: DeviceGroup, ret_type: &syn::ReturnType) -> Tokens {
     quote! { #connect#check }
   };
 
-  // TODO: There should be a better error code returned on the
-  //       `nitrokey` side of things.
+  let connect_err = quote! {
+    ::nitrokey::Error::CommunicationError(::nitrokey::CommunicationError::NotConnected)
+  };
   let skip = if let DeviceGroup::No = group {
-    quote! {let Err(::nitrokey::CommandError::Undefined) = result {} else}
+    quote! {let Err(#connect_err) = result {} else}
   } else {
-    quote! {let Err(::nitrokey::CommandError::Undefined) = result}
+    quote! {let Err(#connect_err) = result}
   };
 
   let result = if let DeviceGroup::No = group {